Initial framework for file listing
The first version of the new structure I'm working on for listing
spatial data. Currently contains a large quantity of  placeholders and
doesn't actually list files. Commintting due to some previous issues
with local files corrupting. Also included a previous working version of
the script.

Also updated spatialTables to be able to list containers and also to
take multiple arguments.